Medical-ChatBot-main/requirements.txt
langchain==0.3.26
langchain-community==0.3.26
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0
pinecone-client==4.1.0
langchain-pinecone==0.2.8

# Google Gemini (used for embeddings only)
langchain-google-genai
google-generativeai

# Groq (used for LLM - higher free-tier limits)
langchain-groq